feat(contract-e2e): auto-source substrate-contracts-node with e2e tests (#254)
* feat(contracts-e2e): auto-source contracts-node binary if not present

* chore(contract-e2e): deprecate --features e2e-tests

* fix: remove unnecessary async function

* refactor(contracts-node): reduce duplicated code with binary sourcing (#255)

* refactor(up-contract): use Binary struct from parachains -> up -> sourcing to auto-launch contracts-node

* refactor(contracts_node): reduce duplicated code -- checkpoint, not working

* refactor(contracts_node): use Binary and Source structs for substrate-contracts-node

* chore: small changes

* chore: small changes

* chore: add e2e help, prevent cmd output, and cleanup

* refactor(contracts_node): introduce helper to download contracts node if it does not exist

* refactor(sourcing): move sourcing functionality to `pop-common` (#258)

* refactor(sourcing): move sourcing to pop-common

* refactor(sourcing-tests): move sourcing tests to pop-common

* refactor(sourcing): better imports

* refactor(sourcing): clean sourcing module with better component categorization. Other cleanups

* fix: failing tests

* fix: needed to download contracts_node for test

* fix: add contracts feature, more tests need s c n downloaded

* fix: manually create temp dir

* fix: append pop_tmp folder in tests

* refactor: e2e test improvements (#263)

* fix: use async sleep in async context

* refactor: remove unnecessary module

* refactor: standardise sourcing ux

* fix: temp dir issue

* chore: component locations, minor improvements

---------
