+++ crates/pop-cli/src/common/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(feature = "contract")]
+pub mod contracts;
