+++ crates/pop-common/src/lib.rs
@@ -3,6 +3,7 @@ pub mod errors;
 pub mod git;
 pub mod helpers;
 pub mod manifest;
+pub mod sourcing;
 pub mod templates;
 
 pub use build::Profile;
@@ -11,3 +12,61 @@ pub use git::{Git, GitHub, Release};
 pub use helpers::{get_project_name_from_path, replace_in_file};
 
 static APP_USER_AGENT: &str = concat!(env!("CARGO_PKG_NAME"), "/", env!("CARGO_PKG_VERSION"));
+
+/// Trait for observing status updates.
+pub trait Status {
+	/// Update the observer with the provided `status`.
+	fn update(&self, status: &str);
+}
+
+impl Status for () {
+	// no-op: status updates are ignored
+	fn update(&self, _: &str) {}
+}
+
+/// Determines the target triple based on the current platform.
+pub fn target() -> Result<&'static str, Error> {
+	use std::env::consts::*;
+
+	if OS == "windows" {
+		return Err(Error::UnsupportedPlatform { arch: ARCH, os: OS });
+	}
+
+	match ARCH {
+		"aarch64" => {
+			return match OS {
+				"macos" => Ok("aarch64-apple-darwin"),
+				_ => Ok("aarch64-unknown-linux-gnu"),
+			}
+		},
+		"x86_64" | "x86" => {
+			return match OS {
+				"macos" => Ok("x86_64-apple-darwin"),
+				_ => Ok("x86_64-unknown-linux-gnu"),
+			}
+		},
+		&_ => {},
+	}
+	Err(Error::UnsupportedPlatform { arch: ARCH, os: OS })
+}
+
+#[cfg(test)]
+mod test {
+	use super::*;
+	use anyhow::Result;
+
+	#[test]
+	fn target_works() -> Result<()> {
+		use std::{process::Command, str};
+		let output = Command::new("rustc").arg("-vV").output()?;
+		let output = str::from_utf8(&output.stdout)?;
+		let target_expected = output
+			.lines()
+			.find(|l| l.starts_with("host: "))
+			.map(|l| &l[6..])
+			.unwrap()
+			.to_string();
+		assert_eq!(target()?, target_expected);
+		Ok(())
+	}
+}
