+++ crates/pop-contracts/Cargo.toml
@@ -1,12 +1,12 @@
 [package]
-name = "pop-contracts"
 description = "Library for generating, building, deploying, and calling ink! smart contracts."
-version = "0.2.0"
-license = "Apache-2.0"
 documentation = "https://docs.rs/pop-contracts/latest/pop_contracts"
 edition.workspace = true
+license = "Apache-2.0"
+name = "pop-contracts"
 readme = "README.md"
 repository.workspace = true
+version = "0.2.0"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
@@ -21,14 +21,14 @@ thiserror.workspace = true
 tokio.workspace = true
 url.workspace = true
 
+heck.workspace = true
 ink_env.workspace = true
 sp-core.workspace = true
 sp-weights.workspace = true
-subxt-signer.workspace = true
-subxt.workspace = true
 strum.workspace = true
 strum_macros.workspace = true
-heck.workspace = true
+subxt-signer.workspace = true
+subxt.workspace = true
 
 # cargo-contracts
 contract-build.workspace = true
@@ -38,5 +38,6 @@ contract-extrinsics.workspace = true
 pop-common = { path = "../pop-common", version = "0.2.0" }
 
 [dev-dependencies]
+dirs.workspace = true
 mockito.workspace = true
 tokio-test.workspace = true
