+++ crates/pop-contracts/src/errors.rs
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: GPL-3.0
 
+use pop_common::sourcing::Error as SourcingError;
 use thiserror::Error;
 
 #[derive(Error, Debug)]
@@ -48,4 +49,6 @@ pub enum Error {
 	UnsupportedPlatform { os: &'static str },
 	#[error("{0}")]
 	UploadContractError(String),
+	#[error("Sourcing error {0}")]
+	SourcingError(SourcingError),
 }
