+++ crates/pop-contracts/src/up.rs
@@ -199,7 +199,10 @@ pub async fn upload_smart_contract(
 #[cfg(test)]
 mod tests {
 	use super::*;
-	use crate::{create_smart_contract, errors::Error, run_contracts_node, templates::Contract};
+	use crate::{
+		create_smart_contract, download_contracts_node, errors::Error, run_contracts_node,
+		templates::Contract,
+	};
 	use anyhow::Result;
 	use std::{env, fs, process::Command};
 	use url::Url;
@@ -335,9 +338,11 @@ mod tests {
 		const LOCALHOST_URL: &str = "ws://127.0.0.1:9944";
 		let temp_dir = generate_smart_contract_test_environment()?;
 		mock_build_process(temp_dir.path().join("testing"))?;
-		// Run contracts-node
-		let cache = temp_dir.path().join("cache");
-		let process = run_contracts_node(cache, None).await?;
+
+		let cache = temp_dir.path().join("");
+
+		let node_path = download_contracts_node(cache.clone()).await?;
+		let process = run_contracts_node(node_path.path(), None).await?;
 
 		let upload_exec = set_up_upload(UpOpts {
 			path: Some(temp_dir.path().join("testing")),
@@ -386,6 +391,7 @@ mod tests {
 			.args(["-s", "TERM", &process.id().to_string()])
 			.spawn()?
 			.wait()?;
+
 		Ok(())
 	}
 }
