+++ crates/pop-contracts/src/utils/mod.rs
@@ -1,5 +1,4 @@
 // SPDX-License-Identifier: GPL-3.0
 
-pub mod contracts_node;
 pub mod helpers;
 pub mod signer;
